+++ src/axom/quest/tests/quest_inout_octree.cpp
@@ -17,20 +17,14 @@
 
 #include "gtest/gtest.h"
 
-#include "axom/core/Types.hpp"
-#include "axom/core/utilities/Timer.hpp"
-
-#include "axom/primal/operators/orientation.hpp"
-#include "axom/primal/geometry/Triangle.hpp"
+#include "axom/core.hpp"
+#include "axom/slic.hpp"
+#include "axom/primal.hpp"
+#include "axom/mint.hpp"
 
 // _quest_inout_cpp_include_start
 #include "axom/quest/geom/InOutOctree.hpp"
 // _quest_inout_cpp_include_end
-
-#include "axom/mint/mesh/Mesh.hpp"
-// #include "axom/mint/utils/vtk_utils.hpp"
-
-#include "axom/slic/interface/slic.hpp"
 
 #include "quest_test_utilities.hpp"
 
@@ -42,14 +36,15 @@ const int DIM = 3;
 }
 
 // _quest_inout_cpp_typedef_start
-typedef axom::quest::InOutOctree<DIM> Octree3D;
+using Octree3D = axom::quest::InOutOctree<DIM>;
 
-typedef Octree3D::GeometricBoundingBox GeometricBoundingBox;
-typedef Octree3D::SpacePt SpacePt;
+using GeometricBoundingBox = Octree3D::GeometricBoundingBox;
+using SpacePt = Octree3D::SpacePt;
 // _quest_inout_cpp_typedef_end
-typedef Octree3D::SpaceVector SpaceVector;
-typedef Octree3D::GridPt GridPt;
-typedef Octree3D::BlockIndex BlockIndex;
+
+using SpaceVector = Octree3D::SpaceVector;
+using GridPt = Octree3D::GridPt;
+using BlockIndex = Octree3D::BlockIndex;
 
 
 #include <cstdlib>
@@ -65,8 +60,27 @@ typedef Octree3D::BlockIndex BlockIndex;
 #endif
 
 
+/// Returns a SpacePt corresponding to the given vertex id \a vIdx  in \a mesh
+SpacePt getVertex(axom::mint::Mesh*& mesh, int vIdx)
+{
+  SpacePt pt;
+  mesh->getNode(vIdx, pt.data());
 
+  return pt;
+}
 
+GeometricBoundingBox computeBoundingBox(axom::mint::Mesh*& mesh)
+{
+  GeometricBoundingBox bbox;
+  for(int i=0 ; i < mesh->getNumberOfNodes() ; ++i)
+  {
+    bbox.addPoint( getVertex(mesh, i) );
+  }
+
+  return bbox;
+}
+
+/// Runs randomized inout queries on an octahedron mesh
 void queryOctahedronMesh(axom::mint::Mesh*& mesh,
                          const GeometricBoundingBox& bbox)
 {
@@ -78,50 +92,67 @@ void queryOctahedronMesh(axom::mint::Mesh*& mesh,
   octree.generateIndex();
   // _quest_inout_cpp_init_end
 
+  SLIC_INFO("Testing point containment on an octahedron surface mesh.");
+  SLIC_INFO("Note: Points on the surface might issue a warning, "
+            << "but should be considered outside the surface.");
+  SLIC_INFO("--[==[");
+
   // Query the mesh containment
   axom::utilities::Timer timer(true);
   for(int i=0 ; i < NUM_PT_TESTS ; ++i)
   {
     SpacePt pt;
 
-//        switch(i)  // test a few special points and a lot of random points
-//        {
-//        case 0: case 1:  case 2:            // Test point at mesh vertices
-//        case 3: case 4:  case 5:
-//            pt = verts[i];
-//            break;
-//        case 6:  case 7:  case 8:  case 9:  // Test point at triangle centers
-//        case 10: case 11: case 12: case 13:
-//        {
-//            int tIdx = (i-6)*VERTS_PER_TRI;
-//            pt = getCentroid( verts[ tvRelation[tIdx]]
-//                            , verts[ tvRelation[tIdx+1]]
-//                            , verts[ tvRelation[tIdx+2]] );
-//        }
-//            break;
-//        case 14: case 15: case 16: case 17:   // Test point at edge centers
-//        case 18: case 19: case 20: case 21:
-//        case 22: case 23: case 24: case 25:
-//        {
-//            int eIdx = (i-14)*VERTS_PER_EDGE;
-//            pt = getCentroid( verts[ evRelation[eIdx]]
-//                            , verts[ evRelation[eIdx +1]] );
-//        }
-//            break;
-//        case 26:                 // origin
-//            pt = SpacePt();
-//            break;
-//        case 27:                 // outside bounding box
-//            pt = SpacePt(2* bbMin);
-//            break;
-//        case 28:                 // outside bounding box
-//            pt = SpacePt(2* bbMax);
-//            break;
-//        default:                // random points in bounding box
-    pt = axom::quest::utilities::randomSpacePt<DIM>(bbMin, bbMax);
-//            break;
-//        }
+    // test a few special points and a lot of random points
+    switch(i)
+    {
+    case 0: case 1:  case 2:                // Test point at mesh vertices
+    case 3: case 4:  case 5:
+      pt = getVertex(mesh, i);
+      break;
+    case 6:  case 7:  case 8:  case 9:      // Test point at triangle centers
+    case 10: case 11: case 12: case 13:
+    {
+      int tIdx = (i-6);
+      GridPt vertInds;
+      mesh->getCellNodeIDs(tIdx, vertInds.data());
 
+      pt = axom::quest::utilities::getCentroid( getVertex(mesh, vertInds[0]),
+                                                getVertex(mesh, vertInds[1]),
+                                                getVertex(mesh, vertInds[2]));
+    }
+    break;
+
+    case 14: case 15: case 16: case 17:     // Test point at edge centers
+    case 18: case 19: case 20: case 21:
+    case 22: case 23: case 24: case 25:
+    {
+      // Define explicit vertex indices for edges of octahedron
+      const int v1[] = {0,0,0,0,1,2,3,4,5,5,5,5};
+      const int v2[] = {1,2,3,4,2,3,4,1,1,2,3,4};
+
+      int eIdx = (i-14);
+      pt = axom::quest::utilities::getCentroid( getVertex(mesh, v1[eIdx]),
+                                                getVertex(mesh, v2[eIdx]));
+    }
+    break;
+
+    case 26:                     // origin
+      pt = SpacePt();
+      break;
+    case 27:                     // outside bounding box
+      pt = SpacePt(2* bbMin);
+      break;
+    case 28:                     // outside bounding box
+      pt = SpacePt(2* bbMax);
+      break;
+    default:                    // random points in bounding box
+      pt = axom::quest::utilities::randomSpacePt<DIM>(bbMin, bbMax);
+      break;
+    }
+
+    // Unit octahedron is a unit sphere under the L1 metric. Points are
+    // inside when the sum of coordinate magnitudes is less than one
     double absCoordSum = std::abs(pt[0]) + std::abs(pt[1]) + std::abs(pt[2]);
 
     // For the time being, we allow the within() test to fail when the
@@ -137,6 +168,7 @@ void queryOctahedronMesh(axom::mint::Mesh*& mesh,
 
   }
   timer.stop();
+  SLIC_INFO("--]==]");
 
   SLIC_INFO("-- querying octahedron with "
             << NUM_PT_TESTS
@@ -177,23 +209,52 @@ TEST( quest_inout_octree, octahedron_mesh)
     << bbox2);
   queryOctahedronMesh(mesh, bbox2);
 
-
   delete mesh;
   mesh = nullptr;
 }
 
 
+TEST( quest_inout_octree, tetrahedron_mesh)
+{
+  SLIC_INFO("*** Exercises InOutOctree queries for several thresholds.\n");
+
+  namespace mint = axom::mint;
+  namespace quest = axom::quest;
+
+  std::vector<double> thresholds = {1E-9, 1E-5, 1E-1, 0.};
+
+  for(auto thresh : thresholds)
+  {
+    mint::Mesh* mesh = quest::utilities::make_tetrahedron_mesh();
+    GeometricBoundingBox bbox = computeBoundingBox(mesh);
+
+    Octree3D octree(bbox, mesh);
+    octree.setVertexWeldThreshold(thresh);
+
+    octree.generateIndex();
+
+    SpacePt queryInside = quest::utilities::getCentroid( getVertex(mesh, 0),
+                                                         getVertex(mesh, 1),
+                                                         getVertex(mesh, 2),
+                                                         getVertex(mesh, 3));
+    SpacePt queryOutside = SpacePt( 2. * bbox.getMax().array() );
+
+    EXPECT_TRUE( octree.within(queryInside) );
+    EXPECT_FALSE( octree.within(queryOutside) );
+
+    delete mesh;
+  }
+}
+
 //----------------------------------------------------------------------
-//----------------------------------------------------------------------
-#include "axom/slic/core/UnitTestLogger.hpp"
-using axom::slic::UnitTestLogger;
 
 int main(int argc, char* argv[])
 {
   ::testing::InitGoogleTest(&argc, argv);
 
-  UnitTestLogger logger;  // create & initialize test logger,
-  axom::slic::setLoggingMsgLevel( axom::slic::message::Debug);
+  namespace slic = axom::slic;
+  slic::UnitTestLogger logger;  // create & initialize test logger,
+  slic::setLoggingMsgLevel( slic::message::Debug);
 
 #ifdef INOUT_OCTREE_TESTER_SHOULD_SEED
   std::srand( std::time(0) );